refactor(lepton): update include directives for consistency and clarity
