+++ shared/include/lepton/core.h
@@ -5,8 +5,8 @@
 
 #include <lepton/core/class.h>
 #include <lepton/core/object.h>
-#include <lepton/result.h>
 #include <lepton/throwable.h>
+#include <lepton/result.h>
 
 RESULTDEF(Size)
 RESULTDEF(BytePtr)
