+++ shared/src/lepton/exception/result_exception.c
@@ -1,4 +1,4 @@
-#include "lepton/class.h"
+#include <lepton/core/class.h>
 #include <lepton/exception/result_exception.h>
 
 INHERITCLASS(ResultException, _RESULT_EXCEPTION_METHODS, EXCEPTION_METHODS)
